changing permission to active user

+++ src/posts/views.py
@@ -1,12 +1,7 @@
 try:
-    from urllib import quote_plus #python 2
+    from urllib import quote_plus 
 except:
     pass
-
-#try:
-   # from urllib.parse import quote_plus #python 3
-#except: 
-   # pass
 
 from django.contrib import messages
 from django.contrib.contenttypes.models import ContentType
@@ -27,8 +22,8 @@ from .models import Post
 def post_create(request):
 	if not request.user.is_authenticated():
 		return redirect("login")
-	#if not request.user.is_staff or not request.user.is_superuser:
-		#raise Http404
+	if not request.user.is_active:
+		raise Http404
 		
 	form = PostForm(request.POST or None, request.FILES or None)
 	if form.is_valid():
@@ -49,8 +44,8 @@ def post_detail(request, slug=None):
 		if not request.user.is_authenticated():
 			return redirect("login")
 
-		#if not request.user.is_staff or not request.user.is_superuser:
-			#raise Http404
+		if not request.user.is_staff or not request.user.is_superuser:
+			raise Http404
 	share_string = quote_plus(instance.content)
 
 	initial_data = {
@@ -137,8 +132,8 @@ def post_list(request):
 def post_update(request, slug=None):
 	if not request.user.is_authenticated():
 		return redirect("login")
-	#if not request.user.is_staff or not request.user.is_superuser:
-		#raise Http404
+	if not request.user.is_active:
+		raise Http404
 	instance = get_object_or_404(Post, slug=slug)
 	form = PostForm(request.POST or None, request.FILES or None, instance=instance)
 	if form.is_valid():
@@ -159,8 +154,8 @@ def post_update(request, slug=None):
 def post_delete(request, slug=None):
 	if not request.user.is_authenticated():
 		return redirect("login")
-	#if not request.user.is_staff or not request.user.is_superuser:
-		#raise Http404
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
 	instance = get_object_or_404(Post, slug=slug)
 	instance.delete()
 	messages.success(request, "Successfully deleted")
